pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "gmp"
version = "0.1.0"
description = "Golden Moment Predictor - 川西旅行景观预测引擎"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "pandas>=2.0",
    "httpx>=0.27",
    "numpy>=1.26",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
